Add extension for binding
