+++ app/src/main/java/org/stepik/android/view/banner/mapper/BannerResourcesMapper.kt
@@ -1,4 +1,4 @@
-package org.stepik.android.view.banner
+package org.stepik.android.view.banner.mapper
 
 import android.content.Context
 import androidx.annotation.ColorInt
